style: Fix clippy lints for casting usize

+++ matchbox_demo/src/box_game.rs
@@ -119,7 +119,7 @@ pub fn setup_scene_system(
         commands
             .spawn_bundle(PbrBundle {
                 mesh: meshes.add(Mesh::from(shape::Cube { size: CUBE_SIZE })),
-                material: materials.add(PLAYER_COLORS[handle as usize].into()),
+                material: materials.add(PLAYER_COLORS[handle].into()),
                 transform,
                 ..default()
             })
@@ -157,7 +157,7 @@ pub fn move_cube_system(
     inputs: Res<Vec<(BoxInput, InputStatus)>>,
 ) {
     for (mut t, mut v, p) in query.iter_mut() {
-        let input = inputs[p.handle as usize].0.inp;
+        let input = inputs[p.handle].0.inp;
         // set velocity through key presses
         if input & INPUT_UP != 0 && input & INPUT_DOWN == 0 {
             v.z -= MOVEMENT_SPEED;
